// src/app/movie-details/store/specific-movie.types.ts
export enum SpecificMovieActionTypes {
  FETCH_DETAILS_START = '[Details Main Component] FETCH_DETAILS_START',
  FETCH_DETAILS_SUCCESS = '[Details Effects] FETCH_DETAILS_SUCCESS',
  FETCH_DETAILS_FAILURE = '[Details Effects] FETCH_DETAILS_FAILURE',
  FETCH_REVIEWS_START = '[Details Main Component] FETCH_REVIEWS_START',
  FETCH_REVIEWS_SUCCESS = '[Details Effects] FETCH_REVIEWS_SUCCESS',
  FETCH_REVIEWS_FAILURE = '[Details Effects] FETCH_REVIEWS_FAILURE',
  FETCH_RECOMMENDED_START = '[Details Main Component] FETCH_RECOMMENDED_START',
  FETCH_RECOMMENDED_SUCCESS = '[Details Effects] FETCH_RECOMMENDED_SUCCESS',
  FETCH_RECOMMENDED_FAILURE = '[Details Effects] FETCH_RECOMMENDED_FAILURE',
  FETCH_SIMILAR_START = '[Details Main Component] FETCH_SIMILAR_START',
  FETCH_SIMILAR_SUCCESS = '[Details Effects] FETCH_SIMILAR_SUCCESS',
  FETCH_SIMILAR_FAILURE = '[Details Effects] FETCH_SIMILAR_FAILURE',
  CLEAR_SPECIFIC_MOVIES = '[Details effects] CLEAR_SPECIFIC_MOVIES',
  UPDATE_SPECIFIC_MOVIE_FAVORITE = '[Details effects] UPDATE_SPECIFIC_MOVIE_FAVORITE',
  UPDATE_SPECIFIC_MOVIE_RATE = '[Details effects] UPDATE_SPECIFIC_MOVIE_RATE',
  UPDATE_SPECIFIC_MOVIE_WATCHLIST_SUCCESS = '[Details effects] UPDATE_SPECIFIC_MOVIE_WATCHLIST_SUCCESS',
  ADD_TO_WATCHLIST_START = '[Movie details movie ]ADD_TO_WATCHLIST_START',
  REMOVE_FROM_WATCHLIST_START = '[Movie details movie ]REMOVE_FROM_WATCHLIST_START',
}

export type SpecificStartActionsTypes =
  | SpecificMovieActionTypes.FETCH_RECOMMENDED_START
  | SpecificMovieActionTypes.FETCH_SIMILAR_START;